requirements.txt
pytorch-lightning
hydra-core
wandb
dvc
python-dotenv
flwr==0.14.0
torchvision
